Move guest cache and anonymous query guards into the plugin
The issues/activities anonymous-access guards and the issues guest action cache are bugs.ruby-lang.org-specific customizations. Relocating them into redmine_bugs_ruby_lang via the plugin's existing class_eval/prepend convention keeps the core Redmine controllers pristine for upstream merges. The Redis cache_store backend stays in config/environments/production.rb since it is app-wide and must be set at environment-config time, before plugins load. No behavior change.
