+++ app/controllers/activities_controller.rb
@@ -20,7 +20,6 @@
 class ActivitiesController < ApplicationController
   menu_item :activity
   before_action :find_optional_project_by_id, :authorize_global
-  before_action :reject_anonymous_activity_filter, :only => :index
   accept_atom_auth :index
 
   def index
@@ -83,17 +82,5 @@ class ActivitiesController < ApplicationController
 
   rescue ActiveRecord::RecordNotFound
     render_404
-  end
-
-  private
-
-  def reject_anonymous_activity_filter
-    return if User.current.logged?
-    render_403 if excessive_activity_params?
-  end
-
-  def excessive_activity_params?
-    params[:from].present? ||
-      params[:user_id].present?
   end
 end
