+++ app/controllers/issues_controller.rb
@@ -20,12 +20,6 @@
 class IssuesController < ApplicationController
   default_search_scope :issues
 
-  caches_action :index,
-    :if => -> { !User.current.logged? && !excessive_query_params? },
-    :expires_in => 5.minutes,
-    :cache_path => -> { request.GET.merge(locale: I18n.locale) }
-
-  before_action :reject_anonymous_issue_filter, :only => :index
   before_action :find_issue, :only => [:show, :edit, :update, :issue_tab]
   before_action :find_issues, :only => [:bulk_edit, :bulk_update, :destroy]
   before_action :authorize, :except => [:index, :new, :create]
@@ -494,19 +488,6 @@ class IssuesController < ApplicationController
   end
 
   private
-
-  def reject_anonymous_issue_filter
-    return if User.current.logged?
-    render_403 if excessive_query_params?
-  end
-
-  def excessive_query_params?
-    params[:set_filter].present? ||
-      params[:query_id].present? ||
-      params[:sort].present? ||
-      params[:per_page].to_i > 50 ||
-      params[:page].to_i > 30
-  end
 
   def query_error(exception)
     session.delete(:issue_query)
