+++ plugins/redmine_bugs_ruby_lang/init.rb
@@ -75,6 +75,9 @@ require 'redmine_mailing_list_integration_imap_supplement/imap'
 require "redmine_ruby_lang_mailing_list_customization/hooks"
 require 'redmine_ruby_lang_mailing_list_customization/redmine_ext'
 
+require 'redmine_bugs_ruby_lang/anonymous_filter/issues_controller'
+require 'redmine_bugs_ruby_lang/anonymous_filter/activities_controller'
+
 ActiveSupport::Inflector.inflections do |inflect|
   inflect.irregular 'use_of_mailng_list', 'uses_of_mailing_list'
 end
